requirements.txt
absl-py==2.1.0
appdirs==1.4.4
asgiref==3.8.1
astunparse==1.6.3
attr==0.3.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
bleach==5.0.1
boto==2.49.0
boto3==1.34.84
botocore==1.34.84
boxing==0.1.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.7.1
Django==3.2.25
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
exceptiongroup==1.2.1
expiringdict==1.2.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gast==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.2
h5py==3.11.0
htmlmin==0.1.12
humansignal-drf-yasg==1.21.9
idna==3.7
ijson==3.2.3
importlib_metadata==7.1.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
jax==0.4.26
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonschema==3.2.0
keras==3.3.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
label-studio==1.11.0
label-studio-converter==0.0.57
label-studio-tools==0.0.3
launchdarkly-server-sdk==7.5.0
libclang==18.1.1
lockfile==0.12.2
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.8
networkx==3.2.1
nltk==3.6.7
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.0.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.5.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.8
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==1.10.15
Pygments==2.17.2
pymycobot==3.4.6
pyparsing==3.1.2
pypng==0.20220715.0
pyRFC3339==1.1
pyrsistent==0.20.0
pyserial==3.5
pytest==8.2.0
python-can==4.3.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
redis==3.5.3
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rq==1.10.1
rsa==4.9
rules==2.2
s3transfer==0.10.1
scipy==1.13.0
seaborn==0.13.2
semver==2.13.0
sentry-sdk==1.45.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.5.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.7.0
tensorflow-cpu==2.16.1
tensorflow-estimator==2.7.0
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ua-parser==0.18.0
ujson==5.9.0
ultralytics==8.2.1
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
xmljson==0.2.0
zipp==3.18.1
